[wip] store vol depth as float

+++ src/aliceVision/depthMap/cuda/planeSweeping/plane_sweeping_cuda.hpp
@@ -12,7 +12,7 @@ namespace aliceVision {
 namespace depthMap {
 
 float ps_planeSweepingGPUPixelsVolume(CudaArray<uchar4, 2>** ps_texs_arr,
-                                      int* ovol_hmh, cameraStruct** cams, int ncams,
+                                      float* ovol_hmh, cameraStruct** cams, int ncams,
                                       int width, int height,
                                       int volStepXY, int volDimX, int volDimY, int volDimZ,
                                       CudaDeviceMemory<float>& depths_dev,
